pyproject.toml
[project]
name = "money"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "dotenv>=0.9.9",
    "gspread>=6.2.1",
    "pandas>=2.3.3",
    "plaid-python>=38.0.0",
    "python-dotenv>=1.2.1",
]
